rename service directories
