+++ LivingLab.Core/DomainServices/DomainService.cs
@@ -1,6 +1,6 @@
 using LivingLab.Core.Interfaces.Services;
 
-namespace LivingLab.Core.Services;
+namespace LivingLab.Core.DomainServices;
 /// <summary>
 /// Domain service implementations belongs here.
 /// Domain service are classes that are responsible for business logic.
