+++ LivingLab.Web/UIServices/Todo/TodoService.cs
@@ -3,7 +3,7 @@ using AutoMapper;
 using LivingLab.Core.Interfaces.Repositories;
 using LivingLab.Web.Models.DTOs.Todo;
 
-namespace LivingLab.Web.Services.Todo;
+namespace LivingLab.Web.UIServices.Todo;
 /// <summary>
 /// This is a UI-specific service so it belongs in the Web project.
 /// It does not contain any business logic and works with UI-specific types (view models and DTOs).
